--- a/physicalplan/binary/scalar.go
+++ b/physicalplan/binary/scalar.go
@@ -13,6 +13,14 @@ import (
 	"github.com/thanos-community/promql-engine/physicalplan/model"
 )
 
+type ScalarSide int
+
+const (
+	ScalarSideBoth ScalarSide = iota
+	ScalarSideLeft
+	ScalarSideRight
+)
+
 // scalarOperator evaluates expressions where one operand is a scalarOperator.
 type scalarOperator struct {
 	seriesOnce sync.Once
@@ -23,17 +31,22 @@ type scalarOperator struct {
 	numberSelector model.VectorOperator
 	next           model.VectorOperator
 	getOperands    getOperandsFunc
+	operandValIdx  int
 	operation      operation
 }
 
-func NewScalar(pool *model.VectorPool, next model.VectorOperator, numberSelector model.VectorOperator, op parser.ItemType, scalarSideLeft bool) (*scalarOperator, error) {
-	binaryOperation, err := newOperation(op, false)
+func NewScalar(pool *model.VectorPool, next model.VectorOperator, numberSelector model.VectorOperator, op parser.ItemType, scalarSide ScalarSide) (*scalarOperator, error) {
+	binaryOperation, err := newOperation(op, scalarSide != ScalarSideBoth)
 	if err != nil {
 		return nil, err
 	}
+	// operandValIdx 0 means to get lhs as the return value
+	// while 1 means to get rhs as the return value.
+	operandValIdx := 0
 	getOperands := getOperandsScalarRight
-	if scalarSideLeft {
+	if scalarSide == ScalarSideLeft {
 		getOperands = getOperandsScalarLeft
+		operandValIdx = 1
 	}
 
 	// Cache the result of the number selector since it
@@ -51,6 +64,7 @@ func NewScalar(pool *model.VectorPool, next model.VectorOperator, numberSelector
 		numberSelector: numberSelector,
 		operation:      binaryOperation,
 		getOperands:    getOperands,
+		operandValIdx:  operandValIdx,
 	}, nil
 }
 
@@ -80,10 +94,16 @@ func (o *scalarOperator) Next(ctx context.Context) ([]model.StepVector, error) {
 	for _, vector := range in {
 		step := o.pool.GetStepVector(vector.T)
 		for i := range vector.Samples {
-			lhs, rhs := o.getOperands(vector, i, o.scalar)
-			val, _ := o.operation(lhs, rhs)
+			operands := o.getOperands(vector, i, o.scalar)
+			val, keep := o.operation(operands, o.operandValIdx)
+			if !keep {
+				continue
+			}
 			step.Samples = append(step.Samples, val)
 			step.SampleIDs = append(step.SampleIDs, vector.SampleIDs[i])
+		}
+		if len(step.Samples) == 0 {
+			continue
 		}
 		out = append(out, step)
 		o.next.GetPool().PutStepVector(vector)
@@ -111,12 +131,12 @@ func (o *scalarOperator) loadSeries(ctx context.Context) error {
 	return nil
 }
 
-type getOperandsFunc func(v model.StepVector, i int, scalar float64) (float64, float64)
+type getOperandsFunc func(v model.StepVector, i int, scalar float64) [2]float64
 
-func getOperandsScalarLeft(v model.StepVector, i int, scalar float64) (float64, float64) {
-	return scalar, v.Samples[i]
+func getOperandsScalarLeft(v model.StepVector, i int, scalar float64) [2]float64 {
+	return [2]float64{scalar, v.Samples[i]}
 }
 
-func getOperandsScalarRight(v model.StepVector, i int, scalar float64) (float64, float64) {
-	return v.Samples[i], scalar
+func getOperandsScalarRight(v model.StepVector, i int, scalar float64) [2]float64 {
+	return [2]float64{v.Samples[i], scalar}
 }