--- a/physicalplan/binary/scalar.go
+++ b/physicalplan/binary/scalar.go
@@ -123,8 +123,10 @@ func (o *scalarOperator) loadSeries(ctx context.Context) error {
 	}
 	series := make([]labels.Labels, len(vectorSeries))
 	for i := range vectorSeries {
-		lbls := labels.NewBuilder(vectorSeries[i]).Del(labels.MetricName).Labels(nil)
-		series[i] = lbls
+		if vectorSeries[i] != nil {
+			lbls := labels.NewBuilder(vectorSeries[i]).Del(labels.MetricName).Labels(nil)
+			series[i] = lbls
+		}
 	}
 
 	o.series = series