--- a/physicalplan/parse/errors.go
+++ b/physicalplan/parse/errors.go
@@ -3,6 +3,17 @@
 
 package parse
 
-import "github.com/efficientgo/core/errors"
+import (
+	"fmt"
+
+	"github.com/efficientgo/core/errors"
+	"github.com/prometheus/prometheus/promql/parser"
+)
 
 var ErrNotSupportedExpr = errors.New("unsupported expression")
+
+func NotSupportedOperationErr(op parser.ItemType) error {
+	t := parser.ItemTypeStr[op]
+	msg := fmt.Sprintf("operation not supported: %s", t)
+	return errors.Wrap(ErrNotSupportedExpr, msg)
+}