Signed-off-by: Ben Ye <benye@amazon.com>

Fixes https://github.com/thanos-community/promql-engine/issues/54