@fpetkovski 2022-10-04
This looks good now. I think we have a small performance regression due to the more complex data types, but I think that's fine for now. The difference is very small:
```
RangeQuery/binary_operation_with_many_to_one-8          23.3ms ± 1%    23.9ms ± 3%  +2.57%  (p=0.000 n=9+10)
RangeQuery/binary_operation_with_vector_and_scalar-8    15.0ms ± 1%    15.4ms ± 2%  +2.10%  (p=0.000 n=9+10)
RangeQuery/unary_negation-8                             13.3ms ± 1%    13.6ms ± 2%  +1.68%  (p=0.001 n=9+10)
```