@fpetkovski 2022-10-04
Also, would you mind adding a benchmark for something like `http_requests_total > 10`?