@yeya24 2022-10-04
```
name                                                   old time/op    new time/op    delta
RangeQuery/vector_selector-10                            27.9ms ± 1%    11.0ms ± 1%  -60.42%  (p=0.000 n=10+10)
RangeQuery/sum-10                                        50.5ms ± 1%     7.2ms ± 2%  -85.81%  (p=0.000 n=9+10)
RangeQuery/sum_by_pod-10                                  135ms ± 1%      11ms ± 1%  -91.98%  (p=0.000 n=10+10)
RangeQuery/rate-10                                       52.0ms ± 1%    17.7ms ± 0%  -65.87%  (p=0.000 n=9+8)
RangeQuery/sum_rate-10                                   70.7ms ± 1%    12.5ms ± 2%  -82.26%  (p=0.000 n=9+10)
RangeQuery/sum_by_rate-10                                 150ms ± 1%      18ms ± 1%  -87.95%  (p=0.000 n=10+8)
RangeQuery/binary_operation_with_one_to_one-10           10.3ms ± 1%     1.4ms ± 0%  -86.68%  (p=0.000 n=10+9)
RangeQuery/binary_operation_with_many_to_one-10           380ms ± 1%      22ms ± 1%  -94.24%  (p=0.000 n=10+10)
RangeQuery/binary_operation_with_vector_and_scalar-10     235ms ± 1%      15ms ± 1%  -93.69%  (p=0.000 n=10+9)
RangeQuery/unary_negation-10                             27.5ms ± 0%    11.7ms ± 3%  -57.55%  (p=0.000 n=9+9)

name                                                   old alloc/op   new alloc/op   delta
RangeQuery/vector_selector-10                            16.5MB ± 0%    26.4MB ± 0%  +59.78%  (p=0.000 n=10+10)
RangeQuery/sum-10                                        5.31MB ± 0%    7.89MB ± 0%  +48.48%  (p=0.000 n=10+9)
RangeQuery/sum_by_pod-10                                 87.6MB ± 0%    17.5MB ± 0%  -80.07%  (p=0.000 n=10+10)
RangeQuery/rate-10                                       17.6MB ± 0%    28.6MB ± 0%  +62.34%  (p=0.000 n=10+10)
RangeQuery/sum_rate-10                                   6.37MB ± 0%   10.35MB ± 0%  +62.56%  (p=0.000 n=10+10)
RangeQuery/sum_by_rate-10                                80.7MB ± 0%    18.7MB ± 0%  -76.87%  (p=0.000 n=10+8)
RangeQuery/binary_operation_with_one_to_one-10           1.77MB ± 0%    2.78MB ± 0%  +57.11%  (p=0.000 n=10+10)
RangeQuery/binary_operation_with_many_to_one-10          63.8MB ± 0%    32.9MB ± 0%  -48.38%  (p=0.000 n=10+10)
RangeQuery/binary_operation_with_vector_and_scalar-10    31.1MB ± 0%    30.4MB ± 0%   -2.50%  (p=0.000 n=10+10)
RangeQuery/unary_negation-10                             17.6MB ± 0%    27.5MB ± 0%  +56.12%  (p=0.000 n=10+9)

name                                                   old allocs/op  new allocs/op  delta
RangeQuery/vector_selector-10                             69.1k ± 0%     75.8k ± 0%   +9.59%  (p=0.000 n=8+10)
RangeQuery/sum-10                                         73.8k ± 0%     69.8k ± 0%   -5.44%  (p=0.000 n=10+9)
RangeQuery/sum_by_pod-10                                   569k ± 0%      156k ± 0%  -72.64%  (p=0.000 n=10+10)
RangeQuery/rate-10                                        78.2k ± 0%    105.5k ± 0%  +34.94%  (p=0.000 n=7+10)
RangeQuery/sum_rate-10                                    82.9k ± 0%     99.8k ± 0%  +20.40%  (p=0.000 n=9+10)
RangeQuery/sum_by_rate-10                                  576k ± 0%      185k ± 0%  -67.88%  (p=0.000 n=10+7)
RangeQuery/binary_operation_with_one_to_one-10            25.7k ± 0%     20.3k ± 0%  -20.94%  (p=0.000 n=10+10)
RangeQuery/binary_operation_with_many_to_one-10            594k ± 0%      126k ± 0%  -78.79%  (p=0.000 n=10+10)
RangeQuery/binary_operation_with_vector_and_scalar-10     84.4k ± 0%     85.7k ± 0%   +1.56%  (p=0.000 n=10+10)
RangeQuery/unary_negation-10                              78.1k ± 0%     84.9k ± 0%   +8.58%  (p=0.000 n=10+9)
➜  promql-engine git:(support-binops) ✗ make benchmark
Benchmarking old engineupport-binops) 
Benchmarking new engine
name                                                   old time/op    new time/op    delta
RangeQuery/vector_selector-10                            28.0ms ± 1%    11.7ms ± 2%  -58.16%  (p=0.000 n=9+9)
RangeQuery/sum-10                                        46.3ms ± 2%     7.6ms ± 3%  -83.64%  (p=0.000 n=10+10)
RangeQuery/sum_by_pod-10                                  131ms ± 3%      12ms ± 9%  -90.82%  (p=0.000 n=10+10)
RangeQuery/rate-10                                       52.6ms ± 1%    18.9ms ± 3%  -63.99%  (p=0.000 n=10+9)
RangeQuery/sum_rate-10                                   67.8ms ± 4%    13.2ms ± 4%  -80.51%  (p=0.000 n=10+10)
RangeQuery/sum_by_rate-10                                 146ms ± 1%      19ms ± 3%  -87.07%  (p=0.000 n=9+9)
RangeQuery/binary_operation_with_one_to_one-10           11.3ms ± 5%     1.4ms ± 2%  -87.28%  (p=0.000 n=10+9)
RangeQuery/binary_operation_with_many_to_one-10           393ms ± 6%      24ms ±14%  -93.99%  (p=0.000 n=9+10)
RangeQuery/binary_operation_with_vector_and_scalar-10     237ms ± 2%      15ms ± 1%  -93.49%  (p=0.000 n=9+9)
RangeQuery/unary_negation-10                             29.1ms ± 1%    12.3ms ± 2%  -57.60%  (p=0.000 n=9+9)
RangeQuery/vector_and_scalar_comparison-10                191ms ± 1%      16ms ± 2%  -91.78%  (p=0.000 n=10+10)

name                                                   old alloc/op   new alloc/op   delta
RangeQuery/vector_selector-10                            16.5MB ± 0%    26.4MB ± 0%  +59.85%  (p=0.000 n=10+10)
RangeQuery/sum-10                                        5.32MB ± 0%    8.03MB ± 2%  +51.10%  (p=0.000 n=10+10)
RangeQuery/sum_by_pod-10                                 87.7MB ± 0%    17.5MB ± 0%  -80.08%  (p=0.000 n=10+10)
RangeQuery/rate-10                                       17.6MB ± 0%    28.6MB ± 0%  +62.60%  (p=0.000 n=10+10)
RangeQuery/sum_rate-10                                   6.37MB ± 0%   10.38MB ± 1%  +62.98%  (p=0.000 n=8+9)
RangeQuery/sum_by_rate-10                                80.7MB ± 0%    18.8MB ± 1%  -76.75%  (p=0.000 n=10+10)
RangeQuery/binary_operation_with_one_to_one-10           1.77MB ± 0%    2.79MB ± 0%  +57.86%  (p=0.000 n=9+9)
RangeQuery/binary_operation_with_many_to_one-10          63.8MB ± 0%    33.0MB ± 0%  -48.36%  (p=0.000 n=10+8)
RangeQuery/binary_operation_with_vector_and_scalar-10    31.1MB ± 0%    30.4MB ± 0%   -2.39%  (p=0.000 n=10+10)
RangeQuery/unary_negation-10                             17.6MB ± 0%    27.5MB ± 0%  +56.36%  (p=0.000 n=10+10)
RangeQuery/vector_and_scalar_comparison-10               29.7MB ± 0%    30.3MB ± 0%   +2.03%  (p=0.000 n=10+10)

name                                                   old allocs/op  new allocs/op  delta
RangeQuery/vector_selector-10                             69.1k ± 0%     75.8k ± 0%   +9.62%  (p=0.000 n=9+10)
RangeQuery/sum-10                                         73.8k ± 0%     69.9k ± 0%   -5.34%  (p=0.000 n=10+10)
RangeQuery/sum_by_pod-10                                   569k ± 0%      156k ± 0%  -72.64%  (p=0.000 n=10+10)
RangeQuery/rate-10                                        78.2k ± 0%    105.5k ± 0%  +34.97%  (p=0.000 n=6+10)
RangeQuery/sum_rate-10                                    82.9k ± 0%     99.8k ± 0%  +20.42%  (p=0.000 n=7+9)
RangeQuery/sum_by_rate-10                                  576k ± 0%      185k ± 0%  -67.87%  (p=0.000 n=10+10)
RangeQuery/binary_operation_with_one_to_one-10            25.7k ± 0%     20.3k ± 0%  -20.88%  (p=0.000 n=10+9)
RangeQuery/binary_operation_with_many_to_one-10            594k ± 0%      126k ± 0%  -78.79%  (p=0.000 n=10+7)
RangeQuery/binary_operation_with_vector_and_scalar-10     84.4k ± 0%     85.7k ± 0%   +1.59%  (p=0.000 n=10+10)
RangeQuery/unary_negation-10                              78.1k ± 0%     84.9k ± 0%   +8.63%  (p=0.000 n=9+10)
RangeQuery/vector_and_scalar_comparison-10                75.3k ± 0%     85.7k ± 0%  +13.82%  (p=0.000 n=10+9)
```

That's the benchmark result.